refactored, including adding helpers and resources
